requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fastapi==0.104.1
fonttools==4.47.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
kiwisolver==1.4.5
loguru==0.7.2
mysql-connector-python==8.2.0
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==12.0
ament_index_python==1.4.0
geometry_msgs==4.2.4
launch_ros==0.19.7
matplotlib==3.9.1
nav2_common==1.1.15
nav2_simple_commander==1.0.0
nav_msgs==4.2.4
numpy==1.21.5
python-dotenv==1.0.1
PyYAML==6.0.1
rclpy==3.3.13
rosidl_parser==3.1.5
sensor_msgs==4.2.4
setuptools==59.6.0
std_msgs==4.2.4
tf2_ros_py==0.25.7
tf_transformations==1.0.0
visualization_msgs==4.2.4